* il OprogStatus.collCount e crossCount li ho resi long long * stampa le collisioni in outputSummaray.
git-svn-id: file:///Users/demichel/SVNREPOS/CODE@6063 fea3b138-1bf5-0310-9da5-cbb00e418fd0

+++ FROM_CVS/trunk/MDsimul/monohs/mono.h
@@ -314,8 +314,8 @@ struct progStatus
   double accrcmz;
   double time;
   int eventCount;
-  int collCount;
-  int crossCount;
+  long long int collCount;
+  long long int crossCount;
   int wallcollCount;
   int PE[PE_POINTS];
   double ENmin;
